Initial commit
Started heavily modifying citysdk.js to make it
compatible with node-js. It is currently set up
in a way that only supports being run on browsers,
but thankfully doesn't look to require too much
tweaking.

I had to disable the indexedDB local caching because
I couldn't figure out how to get it to work with Node
but that would be something cool to have reenabled so
we wouldn't need to make a request for every single
incoming message.

I think I'll eventually look into making a node module
out of this city-sdk/ repo.